Refactor ReadonlyValue
