+++ packages/core/src/common/model/values-test.ts
@@ -1,19 +1,18 @@
 import * as assert from 'assert';
 import {describe, it} from 'mocha';
 
-import {ReadonlyValue} from './readonly-value';
-import {createValue} from './values';
+import {createReadonlyValue, createValue} from './values';
 
-describe(ReadonlyValue.name, () => {
+describe(createReadonlyValue.name, () => {
 	it('should get emitter of value', () => {
 		const v = createValue(0);
-		const [rv] = ReadonlyValue.create(v);
+		const [rv] = createReadonlyValue(v);
 		assert.strictEqual(rv.emitter, v.emitter);
 	});
 
 	it('should get raw value', () => {
 		const v = createValue(123);
-		const [rv] = ReadonlyValue.create(v);
+		const [rv] = createReadonlyValue(v);
 		assert.strictEqual(rv.rawValue, v.rawValue);
 
 		v.rawValue = 456;
@@ -22,7 +21,7 @@ describe(ReadonlyValue.name, () => {
 
 	it('should set raw value', () => {
 		const v = createValue(123);
-		const [rv, setRawValue] = ReadonlyValue.create(v);
+		const [rv, setRawValue] = createReadonlyValue(v);
 		assert.strictEqual(rv.rawValue, v.rawValue);
 
 		setRawValue(456);
