+++ packages/core/src/common/model/values.ts
@@ -1,7 +1,8 @@
 import {Constraint} from '../constraint/constraint';
 import {ComplexValue} from './complex-value';
 import {PrimitiveValue} from './primitive-value';
-import {Value} from './value';
+import {ReadonlyPrimitiveValue} from './readonly-primitive-value';
+import {ReadonlyValue, Value, ValueChangeOptions} from './value';
 
 interface Config<T> {
 	constraint?: Constraint<T>;
@@ -15,4 +16,20 @@ export function createValue<T>(initialValue: T, config?: Config<T>): Value<T> {
 		return new PrimitiveValue(initialValue);
 	}
 	return new ComplexValue(initialValue, config);
+}
+
+export type SetRawValue<T> = (
+	rawValue: T,
+	options?: ValueChangeOptions | undefined,
+) => void;
+
+export function createReadonlyValue<T>(
+	value: Value<T>,
+): [ReadonlyValue<T>, SetRawValue<T>] {
+	return [
+		new ReadonlyPrimitiveValue(value),
+		(rawValue: T, options: ValueChangeOptions | undefined) => {
+			value.setRawValue(rawValue, options);
+		},
+	];
 }
